More cleanup.  Reduce timeout/delay.
