+++ client-src/elements/chromedash-header.js
@@ -157,8 +157,6 @@ export class ChromedashHeader extends LitElement {
 
   connectedCallback() {
     super.connectedCallback();
-    console.log('chromedash-header connectedCallback email: ',
-      this.user?.email || 'none');
 
     // The user sign-in is desktop only.
     if (IS_MOBILE) {
@@ -230,19 +228,15 @@ export class ChromedashHeader extends LitElement {
     signInTestingButton.innerText = 'Sign in as example@chromium.org';
     signInTestingButton.setAttribute('type', 'button');
     signInTestingButton.setAttribute('data-testid', 'dev-mode-sign-in-button');
-    console.log('chromedash-header setup button to login as developer.');
 
     const appComponent = document.querySelector('chromedash-app');
 
     signInTestingButton.addEventListener('click', () => {
-      console.log('login as developer for testing, and replace the url if successful.');
       // POST to '/dev/mock_login' to login as example@chromium.
       fetch('/dev/mock_login', {method: 'POST'})
         .then((response) => {
-          console.log('fetch response ok:', response.ok);
           if (!response.ok) {
             signInTestingButton.style.color = 'red';
-            console.error('About to throw error re failed login.', response);
             throw new Error('Sign in failed! Response:', response);
           }
           // Avoid reloading the page to display with the logged in user.
@@ -252,7 +246,6 @@ export class ChromedashHeader extends LitElement {
         })
         .then(() => {
           setTimeout(() => {
-            console.log('load the home page after login.');
             window.location = '/';
           }, 5000);
         })
@@ -287,7 +280,6 @@ export class ChromedashHeader extends LitElement {
   }
 
   signOut() {
-    console.log('chromedash-header signOut called email:', this.user?.email || 'none');
     window.csClient.signOut().then(() => {
       window.location.reload();
     });
